requirements.txt
# ── CrewAI core ────────────────────────────────────────────────────────────
crewai[anthropic]>=1.14.0

# ── Market simulator ────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
numpy>=1.26.0

# ── HTTP client (tools + signal scanner) ────────────────────────────────────
httpx>=0.27.0

# ── Config + env ────────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
